Added user authorization and field validation

+++ term2/flask-lessons/trello-clone/blueprints/cards_bp.py
@@ -2,7 +2,7 @@ from flask import Blueprint, request
 from flask_jwt_extended import jwt_required, get_jwt_identity
 from setup import db
 from models.card import Card, CardSchema
-from auth import admin_required
+from auth import *
 from blueprints.comments_bp import comments_bp
 
 
@@ -12,7 +12,7 @@ cards_bp = Blueprint('cards', __name__, url_prefix='/cards')
 @cards_bp.route('/')
 @jwt_required()
 def all_cards():
-    admin_required()
+    authorize()
     # select * from cards;
     # stmt = db.select(Card).where(db.or_(Card.status != 'Done', Card.id > 0)).order_by(Card.title.desc())
     stmt = db.select(Card)
@@ -54,11 +54,11 @@ def create_card():
 @cards_bp.route('/<int:id>', methods=['PUT', 'PATCH'])
 @jwt_required()
 def update_card(id):
-    admin_required()
     card_info = CardSchema(exclude=['id', 'date_created']).load(request.json)
     stmt = db.select(Card).filter_by(id=id) # same as .where(Card.id == id)
     card = db.session.scalar(stmt)
     if card:
+        authorize(card.user_id)
         card.title = card_info.get('title', card.title)
         card.description = card_info.get('description', card.description)
         card.status = card_info.get('status', card.status)
@@ -72,10 +72,10 @@ def update_card(id):
 @cards_bp.route('/<int:id>', methods=['DELETE'])
 @jwt_required()
 def delete_card(id):
-    admin_required()
     stmt = db.select(Card).filter_by(id=id)
     card = db.session.scalar(stmt)
     if card:
+        authorize(card.user_id)
         db.session.delete(card)
         db.session.commit()
         return {}, 200
