+++ term2/flask-lessons/trello-clone/blueprints/comments_bp.py
@@ -2,7 +2,7 @@ from flask import Blueprint, request, abort
 from flask_jwt_extended import jwt_required, get_jwt_identity
 from setup import db
 from models.comment import CommentSchema, Comment
-from auth import admin_required
+from auth import *
 
 comments_bp = Blueprint('comments', __name__, url_prefix='/<int:card_id>/comments')
 
@@ -23,11 +23,12 @@ def create_comment(card_id):
 @comments_bp.route('/<int:comment_id>', methods=['PUT', 'PATCH'])
 @jwt_required()
 def update_comment(card_id, comment_id):
-    admin_required()
+    # authorize()
     comment_info = CommentSchema(only=['message']).load(request.json)
     stmt = db.select(Comment).filter_by(id=comment_id) # .where(Comment.id == id)
     comment = db.session.scalar(stmt)
     if comment:
+        authorize(comment.user_id)
         comment.message = comment_info.get('message', comment.message)
         db.session.commit()
         return CommentSchema().dump(comment)
@@ -38,10 +39,11 @@ def update_comment(card_id, comment_id):
 @comments_bp.route('/<int:comment_id>', methods=['DELETE'])
 @jwt_required()
 def delete_comment(card_id, comment_id):
-    admin_required()
+    # authorize()
     stmt = db.select(Comment).filter_by(id=comment_id) # .where(Comment.id == id)
     comment = db.session.scalar(stmt)
     if comment:
+        authorize(comment.user_id)
         db.session.delete(comment)
         db.session.commit()
         return {}, 200
